+++ libraries/lts/include/mcrl2/lts/detail/liblts_scc.h
@@ -67,7 +67,7 @@ class scc_partitioner
     /** \brief Gives the number of bisimulation equivalence classes of the LTS.
      *  \return The number of bisimulation equivalence classes of the LTS.
      */
-    unsigned int num_eq_classes() const;
+    size_t num_eq_classes() const;
 
     /** \brief Gives the equivalence class number of a state.
      *  The equivalence class numbers range from 0 upto (and excluding)
@@ -75,7 +75,7 @@ class scc_partitioner
      *  \param[in] s A state number.
      *  \return The number of the equivalence class to which \e s
      *    belongs. */
-    unsigned int get_eq_class(const unsigned int s) const;
+    size_t get_eq_class(const size_t s) const;
 
     /** \brief Returns whether two states are in the same bisimulation
      *     equivalence class.
@@ -84,12 +84,12 @@ class scc_partitioner
      * \retval true if \e s and \e t are in the same bisimulation
      *    equivalence class;
      * \retval false otherwise. */
-    bool in_same_class(const unsigned int s, const unsigned int t) const;
+    bool in_same_class(const size_t s, const size_t t) const;
 
   private:
 
-    typedef unsigned int state_type;
-    typedef unsigned int label_type;
+    typedef size_t state_type;
+    typedef size_t label_type;
 
     LTS_TYPE &aut;
 
@@ -193,18 +193,18 @@ class scc_partitioner
   }
 
   template < class LTS_TYPE>
-  unsigned int scc_partitioner<LTS_TYPE>::num_eq_classes() const
+  size_t scc_partitioner<LTS_TYPE>::num_eq_classes() const
   { return equivalence_class_index;
   }
 
   template < class LTS_TYPE>
-  unsigned int scc_partitioner<LTS_TYPE>::get_eq_class(const unsigned int s) const
+  size_t scc_partitioner<LTS_TYPE>::get_eq_class(const size_t s) const
   { 
     return block_index_of_a_state[s];
   }
 
   template < class LTS_TYPE>
-  bool scc_partitioner<LTS_TYPE>::in_same_class(const unsigned int s, const unsigned int t) const
+  bool scc_partitioner<LTS_TYPE>::in_same_class(const size_t s, const size_t t) const
   { 
     return get_eq_class(s)==get_eq_class(t);
   }
